Refactor predictive parser
