+++ parser/predictive/parsing_table_test.go
@@ -10,56 +10,6 @@ import (
 	"github.com/moorara/algo/set"
 )
 
-func getTestParsingTables() []*ParsingTable {
-	pt0 := NewParsingTable(
-		[]grammar.Terminal{"+", "*", "(", ")", "id", grammar.Endmarker},
-		[]grammar.NonTerminal{"E", "E′", "T", "T′", "F"},
-	)
-
-	pt0.addProduction("E", "id", &grammar.Production{Head: "E", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("T"), grammar.NonTerminal("E′")}})
-	pt0.addProduction("E", "(", &grammar.Production{Head: "E", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("T"), grammar.NonTerminal("E′")}})
-	pt0.addProduction("E′", "+", &grammar.Production{Head: "E′", Body: grammar.String[grammar.Symbol]{grammar.Terminal("+"), grammar.NonTerminal("T"), grammar.NonTerminal("E′")}})
-	pt0.addProduction("E′", ")", &grammar.Production{Head: "E′", Body: grammar.E})
-	pt0.addProduction("E′", grammar.Endmarker, &grammar.Production{Head: "E′", Body: grammar.E})
-	pt0.addProduction("T", "id", &grammar.Production{Head: "T", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("F"), grammar.NonTerminal("T′")}})
-	pt0.addProduction("T", "(", &grammar.Production{Head: "T", Body: grammar.String[grammar.Symbol]{grammar.NonTerminal("F"), grammar.NonTerminal("T′")}})
-	pt0.addProduction("T′", "+", &grammar.Production{Head: "T′", Body: grammar.E})
-	pt0.addProduction("T′", "*", &grammar.Production{Head: "T′", Body: grammar.String[grammar.Symbol]{grammar.Terminal("*"), grammar.NonTerminal("F"), grammar.NonTerminal("T′")}})
-	pt0.addProduction("T′", ")", &grammar.Production{Head: "T′", Body: grammar.E})
-	pt0.addProduction("T′", grammar.Endmarker, &grammar.Production{Head: "T′", Body: grammar.E})
-	pt0.addProduction("F", "id", &grammar.Production{Head: "F", Body: grammar.String[grammar.Symbol]{grammar.Terminal("id")}})
-	pt0.addProduction("F", "(", &grammar.Production{Head: "F", Body: grammar.String[grammar.Symbol]{grammar.Terminal("("), grammar.NonTerminal("E"), grammar.Terminal(")")}})
-
-	pt0.setSync("E", ")", true)
-	pt0.setSync("E", grammar.Endmarker, true)
-	pt0.setSync("T", "+", true)
-	pt0.setSync("T", ")", true)
-	pt0.setSync("T", grammar.Endmarker, true)
-	pt0.setSync("F", "+", true)
-	pt0.setSync("F", "*", true)
-	pt0.setSync("F", ")", true)
-	pt0.setSync("F", grammar.Endmarker, true)
-
-	pt1 := NewParsingTable(
-		[]grammar.Terminal{"a", "b", "e", "i", "t"},
-		[]grammar.NonTerminal{"S", "S′", "E"},
-	)
-
-	pt1.addProduction("S", "a", &grammar.Production{Head: "S", Body: grammar.String[grammar.Symbol]{grammar.Terminal("a")}})
-	pt1.addProduction("S", "i", &grammar.Production{Head: "S", Body: grammar.String[grammar.Symbol]{grammar.Terminal("i"), grammar.NonTerminal("E"), grammar.Terminal("t"), grammar.NonTerminal("S"), grammar.NonTerminal("S′")}})
-	pt1.addProduction("S′", "e", &grammar.Production{Head: "S′", Body: grammar.E})
-	pt1.addProduction("S′", "e", &grammar.Production{Head: "S′", Body: grammar.String[grammar.Symbol]{grammar.Terminal("e"), grammar.NonTerminal("S")}})
-	pt1.addProduction("S′", grammar.Endmarker, &grammar.Production{Head: "S′", Body: grammar.E})
-	pt1.addProduction("E", "b", &grammar.Production{Head: "E", Body: grammar.String[grammar.Symbol]{grammar.Terminal("b")}})
-
-	pt2 := NewParsingTable(
-		[]grammar.Terminal{"+", "*", "(", ")", "id", grammar.Endmarker},
-		[]grammar.NonTerminal{"E", "T", "F"},
-	)
-
-	return []*ParsingTable{pt0, pt1, pt2}
-}
-
 func TestBuildParsingTable(t *testing.T) {
 	pt := getTestParsingTables()
